Rename base.rb to color.rb
