+++ lib/rb_colors.rb
@@ -1,3 +1,3 @@
 require 'rb_colors/version'
-require 'rb_colors/base'
+require 'rb_colors/color'
 require 'rb_colors/primary'
