+++ lib/rb_colors/color.rb
